+++ udata/core/dataservices/rdf.py
@@ -6,13 +6,14 @@ from udata.core.dataservices.models import HarvestMetadata as HarvestDataservice
 from udata.core.dataset.models import Dataset, License
 from udata.core.dataset.rdf import dataset_to_graph_id, sanitize_html
 from udata.rdf import (
+    CONTACT_POINT_ENTITY_TO_ROLE,
     DCAT,
     DCATAP,
     DCT,
     HVD_LEGISLATION,
     TAG_TO_EU_HVD_CATEGORIES,
-    contact_point_from_rdf,
-    contact_point_to_rdf,
+    contact_points_from_rdf,
+    contact_points_to_rdf,
     namespace_manager,
     rdf_value,
     remote_url_from_rdf,
@@ -43,7 +44,13 @@ def dataservice_from_rdf(
     dataservice.base_api_url = url_from_rdf(d, DCAT.endpointURL)
     dataservice.endpoint_description_url = url_from_rdf(d, DCAT.endpointDescription)
 
-    dataservice.contact_point = contact_point_from_rdf(d, dataservice) or dataservice.contact_point
+    roles = [  # Imbricated list of contact points for each role
+        contact_points_from_rdf(d, rdf_entity, role, dataservice)
+        for rdf_entity, role in CONTACT_POINT_ENTITY_TO_ROLE.items()
+    ]
+    dataservice.contact_points = [  # Flattened list of contact points
+        contact_point for role in roles for contact_point in role
+    ] or dataservice.contact_points
 
     datasets = []
     for dataset_node in d.objects(DCAT.servesDataset):
@@ -156,12 +163,10 @@ def dataservice_to_rdf(dataservice: Dataservice, graph=None):
 
     if is_hvd:
         # We also want to automatically add any HVD category tags of the dataservice's datasets.
-        for dataset in dataservice.datasets:
-            if "hvd" not in dataset.tags:  # Only check HVD datasets for their categories.
-                continue
-            for tag in dataset.tags:
-                if tag in TAG_TO_EU_HVD_CATEGORIES:
-                    hvd_category_tags.add(tag)
+        dataset_ids = [dat.id for dat in dataservice.datasets]
+        for tag in TAG_TO_EU_HVD_CATEGORIES:
+            if Dataset.objects(id__in=dataset_ids, tags="hvd").filter(tags=tag).first():
+                hvd_category_tags.add(tag)
     for tag in hvd_category_tags:
         d.add(DCATAP.hvdCategory, URIRef(TAG_TO_EU_HVD_CATEGORIES[tag]))
 
@@ -178,8 +183,28 @@ def dataservice_to_rdf(dataservice: Dataservice, graph=None):
         for dataset in dataservice.datasets:
             d.add(DCAT.servesDataset, dataset_to_graph_id(dataset))
 
-    contact_point = contact_point_to_rdf(dataservice.contact_point, graph)
-    if contact_point:
-        d.set(DCAT.contactPoint, contact_point)
+    for contact_point, predicate in contact_points_to_rdf(dataservice.contact_points, graph):
+        d.set(predicate, contact_point)
 
     return d
+
+
+def dataservice_as_distribution_to_rdf(
+    dataservice: Dataservice, graph: Graph = None, is_hvd: bool = True
+):
+    """
+    Create a blank distribution pointing towards a dataservice with DCAT.accessService property
+    """
+    id = BNode()
+    distribution = graph.resource(id)
+    distribution.set(RDF.type, DCAT.Distribution)
+    distribution.add(DCT.title, Literal(dataservice.title))
+    distribution.add(DCAT.accessURL, URIRef(dataservice.base_api_url))
+
+    if is_hvd:
+        # DCAT-AP HVD applicable legislation is also expected at the distribution level
+        distribution.add(DCATAP.applicableLegislation, URIRef(HVD_LEGISLATION))
+
+    distribution.add(DCAT.accessService, dataservice_to_rdf(dataservice, graph))
+
+    return distribution