+++ udata/core/dataset/forms.py
@@ -72,7 +72,7 @@ class BaseResourceForm(ModelForm):
         [validators.DataRequired()],
         choices=list(RESOURCE_FILETYPES.items()),
         default="file",
-        description=_("Whether the resource is an uploaded file, " "a remote file or an API"),
+        description=_("Whether the resource is an uploaded file, a remote file or an API"),
     )
     type = fields.RadioField(
         _("Type"),
@@ -89,7 +89,7 @@ class BaseResourceForm(ModelForm):
     checksum = fields.FormField(ChecksumForm)
     mime = fields.StringField(
         _("Mime type"),
-        description=_("The mime type associated to the extension. " "(ex: text/plain)"),
+        description=_("The mime type associated to the extension. (ex: text/plain)"),
     )
     filesize = fields.IntegerField(
         _("Size"), [validators.optional()], description=_("The file size in bytes")
@@ -122,13 +122,13 @@ def validate_contact_point(form, field):
     """Validates contact point with dataset's org or owner"""
     from udata.models import ContactPoint
 
-    if field.data:
+    for contact_point in field.data or []:
         if form.organization.data:
             contact_point = ContactPoint.objects(
-                id=field.data.id, organization=form.organization.data
+                id=contact_point.id, organization=form.organization.data
             ).first()
         elif form.owner.data:
-            contact_point = ContactPoint.objects(id=field.data.id, owner=form.owner.data).first()
+            contact_point = ContactPoint.objects(id=contact_point.id, owner=form.owner.data).first()
         if not contact_point:
             raise validators.ValidationError(
                 _("Wrong contact point id or contact point ownership mismatch")
@@ -145,7 +145,7 @@ class DatasetForm(ModelForm):
     description = fields.MarkdownField(
         _("Description"),
         [validators.DataRequired(), validators.Length(max=DESCRIPTION_SIZE_LIMIT)],
-        description=_("The details about the dataset " "(collection process, specifics...)."),
+        description=_("The details about the dataset (collection process, specifics...)."),
     )
     license = fields.ModelSelectField(_("License"), model=License, allow_blank=True)
     frequency = fields.SelectField(
@@ -168,14 +168,14 @@ class DatasetForm(ModelForm):
     tags = fields.TagField(_("Tags"), description=_("Some taxonomy keywords"))
     private = fields.BooleanField(
         _("Private"),
-        description=_("Restrict the dataset visibility to you or " "your organization only."),
+        description=_("Restrict the dataset visibility to you or your organization only."),
     )
 
     owner = fields.CurrentUserField()
     organization = fields.PublishAsField(_("Publish as"))
     extras = fields.ExtrasField()
     resources = fields.NestedModelList(ResourceForm)
-    contact_point = fields.ContactPointField(validators=[validate_contact_point])
+    contact_points = fields.ContactPointListField(validators=[validate_contact_point])
 
 
 class ResourcesListForm(ModelForm):