+++ udata/core/reuse/models.py
@@ -30,7 +30,7 @@ class ReuseQuerySet(OwnedQuerySet):
         return self(db.Q(private=True) | db.Q(datasets__0__exists=False) | db.Q(deleted__ne=None))
 
 
-def check_url_does_not_exists(url):
+def check_url_does_not_exists(url, **_kwargs):
     """Ensure a reuse URL is not yet registered"""
     if url and Reuse.url_exists(url):
         raise FieldValidationError(_("This URL is already registered"), field="url")
@@ -80,9 +80,9 @@ class Reuse(db.Datetimed, WithMetrics, ReuseBadgeMixin, Owned, db.Document):
         filterable={},
     )
     url = field(
-        db.StringField(required=True),
+        db.URLField(required=True),
         description="The remote URL (website)",
-        check=check_url_does_not_exists,
+        checks=[check_url_does_not_exists],
     )
     urlhash = db.StringField(required=True, unique=True)
     image_url = db.StringField()