+++ udata/tests/api/test_organizations_api.py
@@ -264,8 +264,8 @@ class MembershipAPITest:
         assert len(response.json) == 1
         assert response.json[0]["comment"] == "test"
         assert (
-            response.json[0]["user"]["email"] == "thibaud@example.org"
-        )  # Can see email of applicant
+            response.json[0]["user"]["email"] == "th*****@example.org"
+        )  # Can see partially obfuscated email of applicant
 
     def test_only_org_member_can_get_membership_requests(self, api):
         api.login()
@@ -276,6 +276,42 @@ class MembershipAPITest:
         response = api.get(url_for("api.request_membership", org=organization))
         assert403(response)
 
+    def test_applicant_can_get_their_membership_requests(self, api):
+        applicant = api.login()
+        membership_request = MembershipRequest(user=applicant, comment="test")
+        organization = OrganizationFactory(members=[], requests=[membership_request])
+
+        response = api.get(
+            url_for("api.request_membership", org=organization),
+            query_string={"user": str(applicant.id)},
+        )
+        assert200(response)
+
+    @pytest.mark.parametrize(
+        "searched_status",
+        [
+            "pending",
+            "accepted",
+            "refused",
+        ],
+    )
+    def test_applicant_can_get_their_membership_requests_with_status(
+        self, api, searched_status: str
+    ):
+        applicant = api.login()
+        membership_request = MembershipRequest(user=applicant, comment="test")
+        organization = OrganizationFactory(members=[], requests=[membership_request])
+        response = api.get(
+            url_for("api.request_membership", org=organization),
+            query_string={"user": str(applicant.id), "status": searched_status},
+        )
+        assert200(response)
+        requests = response.json
+        if searched_status == "pending":
+            assert len(requests) == 1
+        else:
+            assert len(requests) == 0
+
     def test_get_members_with_or_without_email(self, api):
         admin = Member(
             user=UserFactory(email="admin@example.org"), role="admin", since="2024-04-14"
@@ -285,7 +321,7 @@ class MembershipAPITest:
 
         organization = OrganizationFactory(members=[admin, editor])
 
-        # Admin can see emails
+        # Organization admin can partially see emails
         api.login(admin.user)
         response = api.get(url_for("api.organization", org=organization))
         assert200(response)
@@ -294,12 +330,12 @@ class MembershipAPITest:
         assert len(members) == 2
         assert members[0]["role"] == "admin"
         assert members[0]["since"] == "2024-04-14T00:00:00+00:00"
-        assert members[0]["user"]["email"] == "admin@example.org"
+        assert members[0]["user"]["email"] == "ad***@example.org"
 
         assert members[1]["role"] == "editor"
-        assert members[1]["user"]["email"] == "editor@example.org"
+        assert members[1]["user"]["email"] == "ed****@example.org"
 
-        # Editor can see emails
+        # Organization editor can partially see emails
         api.login(editor.user)
         response = api.get(url_for("api.organization", org=organization))
         assert200(response)
@@ -308,10 +344,10 @@ class MembershipAPITest:
         assert len(members) == 2
         assert members[0]["role"] == "admin"
         assert members[0]["since"] == "2024-04-14T00:00:00+00:00"
-        assert members[0]["user"]["email"] == "admin@example.org"
+        assert members[0]["user"]["email"] == "ad***@example.org"
 
         assert members[1]["role"] == "editor"
-        assert members[1]["user"]["email"] == "editor@example.org"
+        assert members[1]["user"]["email"] == "ed****@example.org"
 
         # Others cannot see emails
         api.login(other)
@@ -326,6 +362,20 @@ class MembershipAPITest:
 
         assert members[1]["role"] == "editor"
         assert members[1]["user"]["email"] is None
+
+        # Super admin of udata can see emails
+        api.login(AdminFactory())
+        response = api.get(url_for("api.organization", org=organization))
+        assert200(response)
+
+        members = response.json["members"]
+        assert len(members) == 2
+        assert members[0]["role"] == "admin"
+        assert members[0]["since"] == "2024-04-14T00:00:00+00:00"
+        assert members[0]["user"]["email"] == "admin@example.org"
+
+        assert members[1]["role"] == "editor"
+        assert members[1]["user"]["email"] == "editor@example.org"
 
     def test_accept_membership(self, api):
         user = api.login()
@@ -907,6 +957,7 @@ class OrganizationContactPointsAPITest:
             "email": "mooneywayne@cobb-cochran.com",
             "name": "Martin Schultz",
             "organization": str(org.id),
+            "role": "contact",
         }
         response = api.post(url_for("api.contact_points"), data)
         assert201(response)